Update package dependencies and versions
Several package dependencies have been updated to their latest versions in this commit. These include both the addition of new dependencies and updates to existing ones as seen in the package.json and pnpm-look.yaml files. This is part of regular maintenance and ensuring the application has the latest and optimized tools.