Add utility to convert objects to FormData
The new utility file 'object-to-form-data.ts' is created in the 'utils' directory. It contains three functions: objectToFormData, which takes an object and optionally a FormData and namespace and converts the object into FormData; constructFormKey, which helps in creating key using namespace and property; and isObjectNotFile, which checks whether the value is an object and not a file.