Add household members type and update households type in database types
In this commit, the household members type was added to the database types which includes its related row, insert, and update types as well as relationships. In addition, several properties related to location information were removed from the households type, and the "isOneToOne" field for the "public_households_user_id_fkey" relationship in households was updated to false.